feat(three/grid): performance stat + missing dispose() method
