+++ .changeset/long-radios-ask.md
@@ -1,0 +1,5 @@
+---
+"@jolly-pixel/three": minor
+---
+
+Implement performance stats for Grid + Missing dispose() method
