+++ packages/three/examples/scripts/utils/pane.ts
@@ -1,5 +1,5 @@
 // Import Third-party Dependencies
-import { Pane } from "tweakpane";
+import { Pane, type FolderApi } from "tweakpane";
 
 // CONSTANTS
 const kDefaultTitle = "three";
@@ -57,4 +57,52 @@ function currentExample(): string {
   const { pathname } = window.location;
 
   return pathname === "/index.html" ? "/" : pathname;
+}
+
+export type MonitorState = Record<string, number | string>;
+
+export interface MonitorField {
+  label: string;
+  /** Number monitors only; ignored for string values. */
+  format?: (value: number) => string;
+}
+
+export type MonitorFields<TState extends MonitorState> = {
+  [K in keyof TState]?: MonitorField;
+};
+
+/**
+ * Binds the named fields of `state` as read-only rows.
+ * Tweakpane's polling ticker is left off (`interval: 0`): the caller decides
+ * when the values are fresh and calls `folder.refresh()`.
+ */
+export function addMonitors<TState extends MonitorState>(
+  folder: FolderApi,
+  state: TState,
+  fields: MonitorFields<TState>
+): void {
+  for (const [key, field] of Object.entries(fields)) {
+    if (!field) {
+      continue;
+    }
+
+    folder.addBinding(state, key, {
+      readonly: true,
+      interval: 0,
+      label: field.label,
+      format: field.format
+    });
+  }
+}
+
+export function formatCount(
+  value: number
+): string {
+  return Math.round(value).toLocaleString("en-US");
+}
+
+export function formatMilliseconds(
+  value: number
+): string {
+  return `${value.toFixed(1)} ms`;
 }
